Made the base foundation of repo
Created the frontend directory and backend directory as well as some other things that may be useful. In the README youll find the commands I ran to get here. I did not add the main content of the pages because I am still unclear of how we want the user to navigate the page. However, we have the folders where these pages and items would go.